requirements.txt
absl-py==2.3.0
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.9
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
datasets==3.6.0
decorator==4.4.2
decord==0.6.0
deepspeed==0.17.0
diffusers==0.32.0
dill==0.3.8
easydict==1.13
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.115.12
ffmpy==0.6.0
filelock==3.18.0
fonttools==4.59.0
frozenlist==1.6.2
fsspec==2025.3.0
ftfy==6.3.1
func_timeout==4.3.5
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
gradio==5.33.0
gradio_client==1.10.2
groovy==0.1.2
grpcio==1.72.1
h11==0.16.0
hf-xet==1.1.3
hjson==3.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.32.4
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
ipdb==0.13.13
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapy==1.2.4
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.4
multiprocess==0.70.16
networkx==3.4.2
ninja==1.11.1.4
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
OpenEXR==3.3.4
orjson==3.10.18
packaging==25.0
pandas==2.3.0
parso==0.8.5
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
polars==1.35.2
polars-runtime-32==1.35.2
proglog==0.1.12
prompt_toolkit==3.0.52
propcache==0.3.1
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pydantic==2.11.5
pydantic_core==2.33.2
pydub==0.25.1
pyexr==0.5.0
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
ruff==0.11.12
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.0
sentry-sdk==2.34.1
shellingham==1.5.4
simsimd==6.4.7
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.46.2
stringzilla==3.12.5
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tifffile==2025.5.10
timm==1.0.15
tokenizers==0.21.1
tomesd==0.1.3
tomli==2.2.1
tomlkit==0.13.3
torch==2.7.1
torchdiffeq==0.2.5
torchsde==0.2.6
torchvision==0.22.1
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.52.4
triton==3.3.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.3.233
ultralytics-thop==2.0.18
urllib3==2.4.0
uvicorn==0.34.3
wandb==0.21.1
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
